Move parsing logic into Parser.ts
